Renames Collision.{hpp,cpp} to PointStates.{hpp,cpp}
