+++ Testbed/Framework/Test.hpp
@@ -28,7 +28,7 @@
 #include <PlayRho/Common/UnitVec.hpp>
 #include <PlayRho/Common/TypeInfo.hpp>
 
-#include <PlayRho/Collision/Collision.hpp>
+#include <PlayRho/Collision/PointStates.hpp>
 #include <PlayRho/Collision/Distance.hpp>
 #include <PlayRho/Collision/RayCastOutput.hpp>
 #include <PlayRho/Collision/ShapeSeparation.hpp>
